+++ packages/genesys/packages/react-native-components/src/input/NumericInput/NumericInput.tsx
@@ -1,50 +1,28 @@
-import { TextInput } from "react-native";
-import * as Localization from "expo-localization";
 import { NumericInputProps } from "./NumericInput.types";
-import { useControlled } from "@peersyst/react-hooks";
-import formatValue, { decimalRegExp, digitRegExp } from "./utils/formatValue";
-import { useComponentConfig, useMergeDefaultProps } from "@peersyst/react-components-core";
+import {
+    useComponentConfig,
+    useMergeDefaultProps,
+    useNumericInput,
+} from "@peersyst/react-components-core";
+import { TextInput } from "../TextInput";
 
 const NumericInput = (props: NumericInputProps): JSX.Element => {
     const {
         value: valueProp,
         defaultValue,
-        onChangeText,
+        onChange: onChangeProp,
         ...rest
     } = useMergeDefaultProps("NumericInput", props);
     const maxDecimals = useComponentConfig("NumericInput").maxDecimals;
 
-    const [value, setValue] = useControlled(defaultValue || "", valueProp, onChangeText);
+    const { value, onChange } = useNumericInput({
+        defaultValue,
+        value: valueProp,
+        onChange: onChangeProp,
+        maxDecimals,
+    });
 
-    const handleChange = (newValue: string): void => {
-        //Check if the decimal separator is correct
-        if (
-            newValue.endsWith(Localization.digitGroupingSeparator) ||
-            isNaN(Number(newValue.replace(digitRegExp, "").replace(decimalRegExp, ".")))
-        )
-            return;
-        else if (newValue === "") {
-            setValue?.("");
-        } else {
-            const [int, dec] = newValue.split(Localization.decimalSeparator);
-            if (dec && dec.length > maxDecimals) return;
-            const rawInt = int.replace(digitRegExp, "");
-            const rawValue =
-                rawInt +
-                (newValue.includes(Localization.decimalSeparator) ? "." : "") +
-                (dec || "");
-            setValue?.(rawValue);
-        }
-    };
-
-    return (
-        <TextInput
-            value={formatValue(value)}
-            onChangeText={handleChange}
-            keyboardType="numeric"
-            {...rest}
-        />
-    );
+    return <TextInput value={value} onChange={onChange} keyboardType="numeric" {...rest} />;
 };
 
 export default NumericInput;