+++ packages/genesys/packages/react-native-components/src/input/NumericInput/NumericInput.types.ts
@@ -1,3 +1,3 @@
-import { TextInputProps } from "react-native";
+import { TextInputProps } from "../TextInput";
 
 export type NumericInputProps = Omit<TextInputProps, "keyboardType">;