+++ packages/genesys/packages/react-components-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.0.10
+
+[NumericInput] Add base NumericInputComponent. (useNumericInput hook)
+[Config] Add locale to the config. Add useLocale hook
+
 ## 1.0.7
 
 [SelectorGroup] Change selectorgroup types