+++ packages/genesys/packages/react-components/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 3.0.25
+
+[Package.json] Update dependecies: react-components-core
+[Config] Add locale to the config
+
 ## 3.0.2
 
 ### BREAKING