+++ packages/genesys/packages/react-native-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Versions
 
+## 2.0.24
+
+[NumericInput] Make NumericInput reactive to the locale
+[Packege.json] Update react-components-core
+[Config] Add locale to the config
+
 ## 2.0.23
 
 [Image] fix image props